Split Makefile for samples.
