+++ sample/calc_half/Makefile
@@ -1,0 +1,1 @@
+../../build/sample_Makefile
