+++ sample/calc_half/interactive_sample_main.cc
