+++ sample/calc_pi/Makefile
@@ -1,0 +1,1 @@
+../../build/sample_Makefile
