+++ sample/calc_pi/pi_by_leibniz.cc
